pyproject.toml
[project]
name = "ragr"
version = "0.1.0"
description = "Multi-tenant RAG platform"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=4.0.0",
    "anthropic>=0.52.0",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.13.0",
    "fastapi>=0.134.0",
    "httpx>=0.28.0",
    "pgvector>=0.4.2",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.1",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0.47",
    "uvicorn>=0.41.0",
    "voyageai>=0.3.2",
]
